requirements.txt
anyio==4.8.0
aptdaemon==2.0.2
attrs==25.3.0
autocommand==2.2.2
av==14.2.0
babel==2.17.0
bcrypt==4.2.0
blinker==1.9.0
Brlapi==0.8.6
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
command-not-found==0.3
configobj==5.0.9
contourpy==1.3.3
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.4.0
defer==1.0.6
distlib==0.4.0
distro==1.9.0
distro-info==1.14
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.11
inflect==7.3.1
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
keyboard==0.13.5
kiwisolver==1.4.9
language-selector==0.1
lap==0.5.12
launchpadlib==2.1.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libarchive-c==5.1
linkify-it-py==2.0.3
louis==3.33.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
nala==0.16.0
networkx==3.6.1
numpy==2.2.6
oauthlib==3.2.2
olefile==0.47
opencv==4.10.0
opencv-python==4.12.0.88
packaging==26.0
passlib==1.7.4
pemmican==1.0.3
pexpect==4.9.0
picamera2==0.3.23
pidng==4.0.9
piexif==1.1.3
pillow==12.1.0
platformdirs==4.3.7
polars==1.37.1
polars-runtime-32==1.37.1
psutil==7.2.2
ptyprocess==0.7.0
pycairo==1.27.0
pycups==2.0.4
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.10.1
PyOpenGL==3.1.9
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyserial==3.5
python-apt==3.0.0+ubuntu1.1
python-dateutil==2.9.0.post0
python-debian==1.0.1+ubuntu1
python-prctl==1.8.1
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.21.0
scikit-video==1.1.11
scipy==1.17.0
setproctitle==1.3.6
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
simplejpeg==1.7.5
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sympy==1.14.0
systemd-python==235
tomli==2.2.1
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.2
ultralytics==8.4.9
ultralytics-thop==2.0.18
unattended-upgrades==0.1
urllib3==2.6.3
v4l2-python3==0.3.5
virtualenv==20.31.2
wadllib==2.0.0
wheel==0.46.1
xdg==5
xkit==0.0.0
